Move unaligned reads schema validator to custom type validator
