+++ src/encoded/schemas/unaligned_reads.json
@@ -15,33 +15,6 @@
         "submitted_id",
         "uuid"
     ],
-    "oneOf": [
-        {
-            "properties": {
-                "read_pair_number": {
-                    "const": "R2"
-                }
-            },
-            "required": [
-                "paired_with",
-                "read_pair_number"
-            ]
-        },
-        {
-            "properties": {
-                "read_pair_number": {
-                    "not": {
-                        "const": "R2"
-                    }
-                }
-            },
-            "not": {
-                "required": [
-                    "paired_with"
-                ]
-            }
-        }
-    ],
     "additionalProperties": false,
     "mixinProperties": [
         {
@@ -146,6 +119,7 @@
             ]
         },
         "paired_with": {
+            "description": "Link to associated R1 paired-end file, if applicable (For R2 files only).",
             "linkTo": "UnalignedReads"
         }
     }
